+++ server/pyproject.toml
@@ -13,6 +13,7 @@ pymongo = "^4.6.1"
 python-dotenv = "^1.0.0"
 gunicorn = "^21.2.0"
 pylint = "^3.0.3"
+pytest = "^8.0.2"
 
 
 [build-system]
